Tarea 2.2: Crear makefile y dividir el programa

	new file:   .gitignore
	copied:     main.c -> gol.c
	new file:   gol.h
	modified:   main.c
	new file:   makefile
